Rename to torchtitan (#221)
Trying out a full renaming pass from torchtrian -> torchtitan,
including:
1. directory structure
2. all names inside the repo itself.
