+++ setup.py
@@ -1,4 +1,7 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates.
+# This software may be used and distributed according to the terms of the Llama 2 Community License Agreement.
+
+# Copyright (c) Meta Platforms, Inc. and affiliates.
 # All rights reserved.
 #
 # This source code is licensed under the BSD-style license found in the
@@ -13,7 +16,7 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="torchtrain",
+    name="torchtitan",
     version="0.0.1",
     packages=find_packages(),
     install_requires=[
@@ -22,5 +25,5 @@ setup(
     description="Package for training large models using native PyTorch",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
-    url="https://github.com/pytorch-labs/torchtrain",
+    url="https://github.com/pytorch-labs/torchtitan",
 )
