+++ torchtitan/float8_linear.py
@@ -4,9 +4,9 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates.
 # All rights reserved
 
-from torchtrain.config_manager import JobConfig
-from torchtrain.logging_utils import logger
-from torchtrain.models.llama import Transformer
+from torchtitan.config_manager import JobConfig
+from torchtitan.logging_utils import logger
+from torchtitan.models.llama import Transformer
 
 
 def build_fp8_linear(model: Transformer, job_config: JobConfig):
