+++ train_configs/llama_7b.toml
@@ -1,4 +1,4 @@
-# TorchTrain Config.toml
+# torchtitan Config.toml
 [job]
 dump_folder = "./outputs"
 description = "LLaMA 7B training"
@@ -18,7 +18,7 @@ save_tb_folder = "tb"
 name = "llama"
 flavor = "7B"
 norm_type = "fused_rmsnorm" # layernorm / np_layernorm / rmsnorm / fused_rmsnorm
-tokenizer_path = "./torchtrain/datasets/tokenizer/tokenizer.model"
+tokenizer_path = "./torchtitan/datasets/tokenizer/tokenizer.model"
 
 [optimizer]
 name = "AdamW"
